% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{am.trans}
\alias{am.trans}
\title{A function performing MA transformation of data with replicates}
\usage{
am.trans(y)
}
\arguments{
\item{y}{expression data with replicates per each condition}
}
\value{
\item{matrix}{MA matrix}
}
\description{
A function performing MA transformation of data with replicates (for internal use).
}
\details{
For duplicate arrays (x1, x2), M=x1/x2 versus A=x1*x2 can facilitate the
investigation of between-duplicate variability in terms of overall intensity.
}
\examples{
## Generating a replicated simulation dataset
simData <- generateData(n.deg=1000)

## Normalization
normData <- generateData(simData[,-7])

## MA transformation
MAdata4X <- am.trans(normData[,1:3])
MAdata4Y <- am.trans(normData[,4:6])
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

