% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{AVplot}
\alias{AVplot}
\title{A function plotting average intensity versus variance}
\usage{
AVplot(data, avg = "mean", w.value = 0.1, logged = TRUE)
}
\arguments{
\item{data}{expression data matrix}

\item{avg}{measure for average value across the columns of the data matrix (default "mean")}

\item{w.value}{parameter for lowess weight value (default 0.1)}

\item{logged}{logical variable indicating log tranformation (default True)}
}
\value{
\item{plot}{average-variance plot}
}
\description{
A function plotting average intensity versus variance
}
\details{
Average versus variance plot of the data can provide a simple characteristics
of the data. The larger variance, the higher variability.
}
\examples{
## Generating a replicated simulation dataset
data("NBparameter")
simData <- generateData(n.deg=1000)

## Normalization
normData <- generateData(simData[,-7])

## MA transformation
AVplot(normData[,1:3])

}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

