% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{generateData}
\alias{generateData}
\title{A function generating simulation data}
\usage{
generateData(n.gene = 20000, n.cond = 2, n.deg = 0, eff = 1000,
  n.rep = 3, disp = 0.25)
}
\arguments{
\item{n.gene}{the number of genes (default 20000)}

\item{n.cond}{the number of conditions (default 2)}

\item{n.deg}{the number of differentially expressed genes (DEGs) (default 0)}

\item{eff}{count difference of DEGs between conditions  (default 1000)}

\item{n.rep}{the number of replicates in each condition (default 3)}

\item{disp}{dispersion parameter for NB distribution (default 0.25)}
}
\value{
\item{matrix}{simulated datasets}
}
\description{
A function generating simulation data
}
\details{
See the reference paper
}
\examples{
## Generating a replicated simulation dataset
simData <- generateData(n.deg=1000)

## Generating a non-replicated simulation dataset
simData.norep <- generateData(n.deg=1000, n.rep=1)
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}
