% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{lpe.var}
\alias{lpe.var}
\title{A function evaluating LPE variance of data with replicates}
\usage{
lpe.var(y, n.bin = n.bin, df = df)
}
\arguments{
\item{y}{gene expressino whose variance to be estimated}

\item{n.bin}{the number of bins}

\item{df}{degree of freedom in smoothing (arg of smooth.spline())}
}
\value{
\item{matrix}{evaluated LPE}
}
\description{
A function evaluating LPE variance of data with replicates
}
\details{
See the reference paper. See also Jain et al. (2003) Bioinformatics.
}
\examples{
## Generating a replicated simulation dataset
simData <- generateData(n.deg=1000)

## Normalization
normData <- generateData(simData[,-7])

## Evaluating lpe object for each condition with replicates
lpe.x <- lpe.var(normData[,1:3], n.bin=100, df=10)
lpe.y <- lpe.var(normData[,4:6], n.bin=100, df=10)
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

