% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{LPEseq.matrans}
\alias{LPEseq.matrans}
\title{A function performing MA transformation of data without replicate}
\usage{
LPEseq.matrans(dat)
}
\arguments{
\item{dat}{expression data with non-replicate per each condition}
}
\value{
\item{matrix}{MA matrix}
}
\description{
A function performing MA transformation of data without replicate. This is a internal function used in LPEseq.var().
}
\details{
See the reference paper
}
\examples{
## Generating a non-replicated simulation dataset
simData.norep <- generateData(n.deg=1000, n.rep=1)

## Normalization
normData.norep <- LPEseq.normalise(simData.norep[,-3])

## MA transformation
MAdata.norep <- LPEseq.matrans(normData.norep)
}

\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

