% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{LPEseq.normalise}
\alias{LPEseq.normalise}
\title{A function Normalising input data according to their total count values}
\usage{
LPEseq.normalise(expr_dat, method = "mean")
}
\arguments{
\item{expr_dat}{count value matrix (not necessarily to be count matrix; FPKM or RPKM can be applied)}

\item{method}{one of c("mean", "median") for summary of the column sums (default "mean")}
}
\value{
\item{matrix}{log2 transformed normalized matrix}
}
\description{
A function Normalising input data according to their total count values
}
\details{
See the reference paper
}
\examples{
## Generating simulation datasets
simData <- generateData(n.deg=1000, n.rep=5)

## Normalization
normData <- LPEseq.normalise(simData[,-11])
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

