% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{LPEseq.outlier}
\alias{LPEseq.outlier}
\title{A function removing outliers}
\usage{
LPEseq.outlier(expr_dat, d = d)
}
\arguments{
\item{expr_dat}{expression data with non-replicate per each condition}

\item{d}{expression difference between condition (log2 scaled)}
}
\value{
\item{matrix}{intput matrix outlier-removed}
}
\description{
A function removing outliers for evaluating non-repliated RNA-seq data
}
\details{
See the reference paper
}
\examples{
## Generating a non-replicated simulation dataset
simData.norep <- generateData(n.deg=1000, n.rep=1)

## Normalization
normData.norep <- LPEseq.normalise(simData.norep[,-3])

## Removing outliers
normData.inuse <- LPEseq.outlier(normData.norep, d = 1.2)
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

