% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{LPEseq.test}
\alias{LPEseq.test}
\title{A function testing differential expression between two conditions}
\usage{
LPEseq.test(expr_x, expr_y, n.bin = 100, df = 10, d = 1.2)
}
\arguments{
\item{expr_x}{a numeric vector or matrix in first condition}

\item{expr_y}{a numeric vector or matrix in second condition}

\item{n.bin}{the number of quantile bins (default 100)}

\item{df}{degrees of freedom in smooth.spline() function (default 10)}

\item{d}{expression difference in LPEseq.outlier() function (defalut 0.5; use 1.2 for biological replicates)}
}
\value{
\item{dataframe}{dataframe containing mean of each condition, local pooled variance, statistics, p-value, and BH adjusted p-value}
}
\description{
A function testing differential expression between two conditions
}
\details{
See the reference paper
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, 2014 (http://bibs.snu.ac.kr/software/LPEseq)
}

