% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{LPEseq.var}
\alias{LPEseq.var}
\title{A function evaluating local pooled error variance with non-replicated data}
\usage{
LPEseq.var(y, n.bin = n.bin, df = df, d = d)
}
\arguments{
\item{y}{expression data with non-replicate per each condition}

\item{n.bin}{the number of quantile bins}

\item{df}{degrees of freedom of smooth.spline() function}

\item{d}{expression difference of LPEseq.outlier() function}
}
\value{
\item{list}{output of smooth.spline() function}
}
\description{
A function evaluating local pooled error variance with non-replicated data
}
\details{
See the reference paper
}
\examples{
## Generating a non-replicated simulation dataset
simData.norep <- generateData(n.deg=1000, n.rep=1)

## Normalization
normData.norep <- LPEseq.normalise(simData.norep[,-3])

## Generating LPE object
normData.var <- LPEseq.var(normData.norep, n.bin=100, df=10, d=1.2)
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

