% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LPEseq.R
\name{permute}
\alias{permute}
\title{A function performing permutation used in am.trans function}
\usage{
permute(a)
}
\arguments{
\item{a}{numeric value indicating the number of samples}
}
\value{
\item{numeric}{permuted number}
}
\description{
A function performing permutation used in am.trans function
}
\details{
See the reference paper
}
\author{
Jungsoo Gim
}
\references{
J. Gim and T. Park, PLoS ONE under revision (http://bibs.snu.ac.kr/software/LPEseq)
}

